chore(mobile): split activity_page below the 1000-line file ceiling
activity_page.dart had grown to 1,072 lines, over the hard 1,000-line
ceiling enforced by mobile/scripts/check-file-sizes.mjs. Move the private
sub-widgets into sibling part files under activity_page/, following the
compose_bar/ pattern:

- header_actions.dart — _FilterMenuButton, _CountBadge, _InboxOptionsButton
- inbox_row.dart — _NewBoundaryDivider, _InboxRow, _RowAvatar
- lists.dart — _RemindersList, _DraftsList, _inboxTimestamp
- status_views.dart — _LoadingSkeleton, _EmptySurface, _EmptyFilterState,
  _ErrorView

Pure code movement — no behavior change, no override added. dart format
reports no changes; flutter analyze and the size guard pass.
